requirements.txt
langchain
langgraph
qdrant-client
torch
transformers
openai
typing-extensions
python-dotenv
